Move redudant check_ip_addr method to common ce (#25575)
