+++ lib/ansible/modules/network/cloudengine/ce_aaa_server_host.py
@@ -282,12 +282,9 @@ updates:
              "hwtacacs server authorization 10.135.182.157 vpn-instance test_vpn public-net"]
 '''
 
-import sys
-import socket
 from xml.etree import ElementTree
 from ansible.module_utils.basic import AnsibleModule
-from ansible.module_utils.ce import get_nc_config, set_nc_config, ce_argument_spec
-
+from ansible.module_utils.ce import get_nc_config, set_nc_config, ce_argument_spec, check_ip_addr
 
 SUCCESS = """success"""
 FAILED = """failed"""
@@ -2191,25 +2188,6 @@ class AaaServerHost(object):
 
         cmds.append(cmd)
         return cmds
-
-
-def check_ip_addr(ipaddr):
-    """ check_ip_addr, Supports IPv4 and IPv6 """
-
-    if not ipaddr or '\x00' in ipaddr:
-        return False
-
-    try:
-        res = socket.getaddrinfo(ipaddr, 0, socket.AF_UNSPEC,
-                                 socket.SOCK_STREAM,
-                                 0, socket.AI_NUMERICHOST)
-        return bool(res)
-    except socket.gaierror:
-        err = sys.exc_info()[1]
-        if err.args[0] == socket.EAI_NONAME:
-            return False
-        raise
-    return True
 
 
 def check_name(**kwargs):
