+++ lib/ansible/modules/network/cloudengine/ce_acl.py
@@ -210,12 +210,9 @@ updates:
     sample: ["undo acl name test"]
 '''
 
-import socket
-import sys
 from xml.etree import ElementTree
 from ansible.module_utils.basic import AnsibleModule
-from ansible.module_utils.ce import get_nc_config, set_nc_config, ce_argument_spec
-
+from ansible.module_utils.ce import get_nc_config, set_nc_config, ce_argument_spec, check_ip_addr
 
 # get acl
 CE_GET_ACL_HEADER = """
@@ -317,25 +314,6 @@ CE_DELETE_ACL_BASE_RULE_TAIL = """
       </acl>
     </config>
 """
-
-
-def check_ip_addr(ipaddr):
-    """ check_ip_addr, Supports IPv4 and IPv6 """
-
-    if not ipaddr or '\x00' in ipaddr:
-        return False
-
-    try:
-        res = socket.getaddrinfo(ipaddr, 0, socket.AF_UNSPEC,
-                                 socket.SOCK_STREAM,
-                                 0, socket.AI_NUMERICHOST)
-        return bool(res)
-    except socket.gaierror:
-        err = sys.exc_info()[1]
-        if err.args[0] == socket.EAI_NONAME:
-            return False
-        raise
-    return True
 
 
 class BaseAcl(object):
